requirements.txt
yfinance==0.2.32
pandas==2.0.3
numpy==1.24.3
python-dateutil==2.8.2
requests==2.31.0
pytest==7.4.0
pytest-cov==4.1.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.6
pydantic==2.5.0
